code/requirements.txt
streamlit==1.32.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv==1.0.1
openai>=1.68.2
chromadb>=0.3.0
tiktoken>=0.3.0
plotly==5.19.0
pandas==2.2.1
numpy==1.26.4
scikit-learn>=1.0.0
nltk>=3.6.0
google-generativeai==0.3.2
pydantic==2.6.3
python-dateutil==2.8.2

# Recommendation Service Dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
plotly>=5.13.0
streamlit>=1.22.0

# Data Processing
datasets>=3.4.1 